chore: export svg types and types array
